+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
+## 2024-05-30
+
+### Changed
+
+- Removed `lust` from bad word list.
+
+### Fixed
+
+- Bad word filter now properly handles words that should be partial matches and words that should not.
+- Updated description for Dying Breath mod to match in-game.
+
 ## 2024-05-29
 
 ### Changed